+++ apps/server-doj/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gsa-tts/forms-server-doj
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @gsa-tts/forms-infra-core@0.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gsa-tts/forms-infra-core@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes